python/Drive/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDrive2==1.15.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.0.2
